ci: skip copyright-lint from GHA
Pre-commit action defaults to run with '--all-files', which is great for
the other action. For copyright-lint however, it is wrong to assume that
all files were changed by a single entity and therefore this can't be
tested easily. It could work if a range of files were passed and the
author email extracted from each commit.